requirements.txt
openai
gradio
python-dotenv
requests
pypdf